Split out different parts of plugin;
