+++ packages/kasia-plugin-wp-api-menus/.babelrc
@@ -1,4 +1,3 @@
 {
-  "presets": ["es2015", "stage-0", "react"],
-  "plugins": ["transform-decorators-legacy"]
+  "presets": ["es2015", "stage-0", "react"]
 }
