+++ packages/kasia-plugin-wp-api-menus/package.json
@@ -11,7 +11,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git@gitlab.outlandish.com:pepperoni/pepperoni.git"
+    "url": "git@gitlab.outlandish.com:pepperoni/pepperoni-plugin-wp-api-menus.git"
   },
   "keywords": [
     "javascript",
