+++ packages/kasia-plugin-wp-api-menus/src/ActionTypes.js
@@ -1,0 +1,7 @@
+export default {
+  REQUEST_MENUS: 'pepperoni/REQUEST_MENUS',
+  REQUEST_MENU: 'pepperoni/REQUEST_MENU',
+  REQUEST_LOCATIONS: 'pepperoni/REQUEST_LOCATIONS',
+  REQUEST_LOCATION: 'pepperoni/REQUEST_LOCATION',
+  RECEIVE_DATA: 'pepperoni/RECEIVE_DATA'
+}
